rename test file to original name
